+++ components/image-cropper/lib/OkImageCropperType.ts
@@ -1,0 +1,7 @@
+export enum OkImageCropperType {
+    // user profile editing modal
+    profile,
+
+    // community creation and update modal
+    community
+}
