+++ services/Apis/communities/CommunitiesApiService.ts
@@ -35,7 +35,8 @@ import {
     BanCommunityUserApiParams,
     UnbanCommunityUserApiParams,
     AddCommunityAdministratorApiParams,
-    AddCommunityModeratorApiParams
+    AddCommunityModeratorApiParams,
+    CreateCommunityApiParams
 } from '~/services/Apis/communities/CommunitiesApiServiceTypes';
 import { AxiosResponse } from '~/node_modules/axios';
 import { CommunityData } from '~/types/models-data/communities/CommunityData';
@@ -491,6 +492,47 @@ export class CommunitiesApiService implements ICommunitiesApiService {
     unfavoriteCommunity(params: UnfavoriteCommunityApiParams): Promise<AxiosResponse<void>> {
         const path = this.makeFavoriteCommunityPath(params.communityName);
         return this.httpService.delete(path, {
+            appendAuthorizationToken: true,
+            isApiRequest: true
+        });
+    }
+
+    createCommunity(params: CreateCommunityApiParams): Promise<AxiosResponse<CommunityData>> {
+        const bodyFormData = new FormData();
+
+        bodyFormData.set('name', params.name);
+        bodyFormData.set('title', params.title);
+        bodyFormData.set('type', params.type.toString());
+        bodyFormData.set('categories', params.categories.join(','));
+
+        if (params.description !== undefined) {
+            bodyFormData.set('description', params.description);
+        }
+
+        if (params.avatar instanceof File || params.avatar instanceof Blob) {
+            bodyFormData.set('avatar', params.avatar, 'avatar.png');
+        }
+
+        if (params.cover instanceof File || params.cover instanceof Blob) {
+            bodyFormData.set('cover', params.cover, 'cover.png');
+        }
+
+        if (params.rules !== undefined) {
+            bodyFormData.set('rules', params.rules);
+        }
+
+        bodyFormData.set('user_adjective', params.userAdjective || 'member');
+        bodyFormData.set('users_adjective', params.usersAdjective || 'members');
+
+        if (params.color) {
+            bodyFormData.set('color', params.color.hex());
+        }
+
+        if (typeof params.invitesEnabled !== 'undefined') {
+            bodyFormData.set('invites_enabled', params.invitesEnabled ? '1' : '0');
+        }
+
+        return this.httpService.put(CommunitiesApiService.CREATE_COMMUNITY_PATH, bodyFormData, {
             appendAuthorizationToken: true,
             isApiRequest: true
         });
