+++ services/navigation/INavigationService.ts
@@ -2,6 +2,7 @@ import VueRouter from 'vue-router';
 import { Context } from '@nuxt/types'
 import { IPost } from '~/models/posts/post/IPost';
 import { IUser } from '~/models/auth/user/IUser';
+import { ICommunity } from '~/models/communities/community/ICommunity';
 
 export interface INavigationService {
     navigateToLogin(config?: NavigationConfig): Promise<void>;
@@ -14,6 +15,8 @@ export interface INavigationService {
 
     navigateToProfile(config?: NavigateToProfileConfig): Promise<void>;
 
+    navigateToCommunity(confir?: NavigateToCommunityConfig): Promise<void>;
+
     setVueRouter(vueRouter: VueRouter): void;
 }
 
@@ -23,6 +26,10 @@ export interface NavigateToPostConfig extends NavigationConfig {
 
 export interface NavigateToProfileConfig extends NavigationConfig {
     user: IUser;
+}
+
+export interface NavigateToCommunityConfig extends NavigationConfig {
+    community: ICommunity;
 }
 
 export interface NavigationConfig {
